+++ src/partykit/gateway.ts
@@ -237,3 +237,17 @@ export class PartyKitTestStore implements bs.TestGateway {
         return buffer.Ok();
     }
 }
+
+export function registerPartyKitStoreProtocol(protocol = "partykit:", overrideBaseURL?: string) {
+    return bs.registerStoreProtocol({
+        protocol,
+        overrideBaseURL,
+        gateway: async (logger) => {
+            return new PartyKitGateway(logger);
+        },
+        test: async (sthis: SuperThis) => {
+            const gateway = new PartyKitGateway(sthis);
+            return new PartyKitTestStore(gateway, sthis);
+        },
+    });
+}
